abstract image provider aliases
