+++ cellprofiler_core/image/abstract_image_provider/__init__.py
@@ -1,5 +1,6 @@
 from ._abstract_image_provider import AbstractImageProvider
-
+from ._callback_image_provider import CallbackImageProvider
+from ._vanilla_image_provider import VanillaImageProvider
 from .load_images_image_provider import LoadImagesFlexFrameProvider
 from .load_images_image_provider import LoadImagesImageProvider
 from .load_images_image_provider import LoadImagesImageProviderURL
