+++ cellprofiler_core/modules/__init__.py
@@ -1,4 +1,6 @@
 # coding=utf-8
+import shutil
+import tempfile
 
 import glob
 import logging
@@ -8,8 +10,13 @@ import re
 import sys
 import types
 
+import six
+
+import cellprofiler_core.measurement
 import cellprofiler_core.module
 import cellprofiler_core.preferences
+from cellprofiler_core.utilities import generate_presigned_url
+from cellprofiler_core.utilities.pathname import url2pathname
 
 logger = logging.getLogger(__name__)
 
@@ -320,3 +327,232 @@ def convert_image_to_objects(image):
     image = numpy.zeros(image.shape[:2], index.dtype)
     image.ravel()[order] = index
     return image
+
+
+UIC1_TAG = 33628
+UIC2_TAG = 33629
+UIC3_TAG = 33630
+UIC4_TAG = 33631
+C_MD5_DIGEST = "MD5Digest"
+C_SCALING = "Scaling"
+C_HEIGHT = "Height"
+C_WIDTH = "Width"
+MS_EXACT_MATCH = "Text-Exact match"
+MS_REGEXP = "Text-Regular expressions"
+MS_ORDER = "Order"
+FF_INDIVIDUAL_IMAGES = "individual images"
+FF_STK_MOVIES = "stk movies"
+FF_AVI_MOVIES = "avi,mov movies"
+FF_AVI_MOVIES_OLD = ["avi movies"]
+FF_OTHER_MOVIES = "tif,tiff,flex,zvi movies"
+FF_OTHER_MOVIES_OLD = ["tif,tiff,flex movies", "tif,tiff,flex movies, zvi movies"]
+IO_IMAGES = "Images"
+IO_OBJECTS = "Objects"
+IO_ALL = (IO_IMAGES, IO_OBJECTS)
+IMAGE_FOR_OBJECTS_F = "IMAGE_FOR_%s"
+SUPPORTED_IMAGE_EXTENSIONS = {
+    ".ppm",
+    ".grib",
+    ".im",
+    ".rgba",
+    ".rgb",
+    ".pcd",
+    ".h5",
+    ".jpe",
+    ".jfif",
+    ".jpg",
+    ".fli",
+    ".sgi",
+    ".gbr",
+    ".pcx",
+    ".mpeg",
+    ".jpeg",
+    ".ps",
+    ".flc",
+    ".tif",
+    ".hdf",
+    ".icns",
+    ".gif",
+    ".palm",
+    ".mpg",
+    ".fits",
+    ".pgm",
+    ".mic",
+    ".fit",
+    ".xbm",
+    ".eps",
+    ".emf",
+    ".dcx",
+    ".bmp",
+    ".bw",
+    ".pbm",
+    ".dib",
+    ".ras",
+    ".cur",
+    ".fpx",
+    ".png",
+    ".msp",
+    ".iim",
+    ".wmf",
+    ".tga",
+    ".bufr",
+    ".ico",
+    ".psd",
+    ".xpm",
+    ".arg",
+    ".pdf",
+    ".tiff",
+}
+SUPPORTED_MOVIE_EXTENSIONS = {
+    ".avi",
+    ".mpeg",
+    ".stk",
+    ".flex",
+    ".mov",
+    ".tif",
+    ".tiff",
+    ".zvi",
+}
+FF = [FF_INDIVIDUAL_IMAGES, FF_STK_MOVIES, FF_AVI_MOVIES, FF_OTHER_MOVIES]
+M_NONE = "None"
+M_FILE_NAME = "File name"
+M_PATH = "Path"
+M_BOTH = "Both"
+M_Z = "Z"
+M_T = "T"
+C_SERIES = "Series"
+C_FRAME = "Frame"
+P_IMAGES = "LoadImagesImageProvider"
+V_IMAGES = 1
+P_MOVIES = "LoadImagesMovieProvider"
+V_MOVIES = 2
+P_FLEX = "LoadImagesFlexFrameProvider"
+V_FLEX = 1
+I_INTERLEAVED = "Interleaved"
+I_SEPARATED = "Separated"
+SUB_NONE = "None"
+SUB_ALL = "All"
+SUB_SOME = "Some"
+
+
+def default_cpimage_name(index):
+    # the usual suspects
+    names = ["DNA", "Actin", "Protein"]
+    if index < len(names):
+        return names[index]
+    return "Channel%d" % (index + 1)
+
+
+def well_metadata_tokens(tokens):
+    """Return the well row and well column tokens out of a set of metadata tokens"""
+
+    well_row_token = None
+    well_column_token = None
+    for token in tokens:
+        if cellprofiler_core.measurement.is_well_row_token(token):
+            well_row_token = token
+        if cellprofiler_core.measurement.is_well_column_token(token):
+            well_column_token = token
+    return well_row_token, well_column_token
+
+
+def needs_well_metadata(tokens):
+    """Return true if, based on a set of metadata tokens, we need a well token
+
+    Check for a row and column token and the absence of the well token.
+    """
+    if cellprofiler_core.measurement.FTR_WELL.lower() in [x.lower() for x in tokens]:
+        return False
+    well_row_token, well_column_token = well_metadata_tokens(tokens)
+    return (well_row_token is not None) and (well_column_token is not None)
+
+
+def is_image(filename):
+    """Determine if a filename is a potential image file based on extension"""
+    ext = os.path.splitext(filename)[1].lower()
+    return ext in SUPPORTED_IMAGE_EXTENSIONS
+
+
+def is_movie(filename):
+    ext = os.path.splitext(filename)[1].lower()
+    return ext in SUPPORTED_MOVIE_EXTENSIONS
+
+
+def is_numpy_file(filename):
+    return os.path.splitext(filename)[-1].lower() == ".npy"
+
+
+def is_matlab_file(filename):
+    return os.path.splitext(filename)[-1].lower() == ".mat"
+
+
+def loadmat(path):
+    imgdata = scipy.io.matlab.mio.loadmat(path, struct_as_record=True)
+    img = imgdata["Image"]
+
+    return img
+
+
+def load_data_file(pathname_or_url, load_fn):
+    ext = os.path.splitext(pathname_or_url)[-1].lower()
+
+    if any([pathname_or_url.startswith(scheme) for scheme in PASSTHROUGH_SCHEMES]):
+        url = generate_presigned_url(pathname_or_url)
+
+        try:
+            src = six.moves.urllib.urlopen(url)
+            fd, path = tempfile.mkstemp(suffix=ext)
+            with os.fdopen(fd, mode="wb") as dest:
+                shutil.copyfileobj(src, dest)
+            img = load_fn(path)
+        finally:
+            try:
+                src.close()
+                os.remove(path)
+            except NameError:
+                pass
+
+        return img
+
+    return load_fn(pathname_or_url)
+
+
+FILE_SCHEME = "file:"
+PASSTHROUGH_SCHEMES = ("http", "https", "ftp", "omero", "s3")
+
+
+def is_file_url(url):
+    return url.lower().startswith(FILE_SCHEME)
+
+
+def urlfilename(url):
+    """Return just the file part of a URL
+
+    For instance http://nucleus.org/linked_files/file%20has%20spaces.txt
+    has a file part of "file has spaces.txt"
+    """
+    if is_file_url(url):
+        return os.path.split(url2pathname(url))[1]
+    path = six.moves.urllib.parse.urlparse(url)[2]
+    if "/" in path:
+        return six.moves.urllib.unquote(path.rsplit("/", 1)[1])
+    else:
+        return six.moves.urllib.unquote(path)
+
+
+def urlpathname(url):
+    """Return the path part of a URL
+
+    For instance, http://nucleus.org/Comma%2Cseparated/foo.txt
+    has a path of http://nucleus.org/Comma,separated
+
+    A file url has the normal sort of path that you'd expect.
+    """
+    if is_file_url(url):
+        return os.path.split(url2pathname(url))[0]
+    scheme, netloc, path = six.moves.urllib.parse.urlparse(url)[:3]
+    path = six.moves.urllib.parse.urlunparse([scheme, netloc, path, "", "", ""])
+    if "/" in path:
+        return six.moves.urllib.unquote(path.rsplit("/", 1)[0])
+    else:
+        return six.moves.urllib.unquote(path)
