+++ tests/image/test_image_set_list.py
@@ -1,6 +1,7 @@
 import numpy
 
 import cellprofiler_core.image
+import cellprofiler_core.image.abstract_image_provider._callback_image_provider
 
 
 class TestImageSetList:
@@ -53,7 +54,7 @@ class TestImageSetList:
             ), "Callback was not called with the correct image provider"
             return img
 
-        z = cellprofiler_core.image.CallbackImageProvider("TestImageProvider", fn)
+        z = cellprofiler_core.image.abstract_image_provider._callback_image_provider.CallbackImageProvider("TestImageProvider", fn)
         y.providers.append(z)
         assert img == y.get_image("TestImageProvider")
 
@@ -72,8 +73,10 @@ class TestImageSetList:
             assert y == image_set, "Callback was not called with the correct image set"
             return img2
 
-        y.providers.append(cellprofiler_core.image.CallbackImageProvider("IP1", fn1))
-        y.providers.append(cellprofiler_core.image.CallbackImageProvider("IP2", fn2))
+        y.providers.append(
+            cellprofiler_core.image.abstract_image_provider._callback_image_provider.CallbackImageProvider("IP1", fn1))
+        y.providers.append(
+            cellprofiler_core.image.abstract_image_provider._callback_image_provider.CallbackImageProvider("IP2", fn2))
         assert img1 == y.get_image("IP1"), "Failed to get correct first image"
         assert img2 == y.get_image("IP2"), "Failed to get correct second image"
